abstract(post fetching functions)
